pyproject.toml
[project]
name = "marketplaats-bot"
version = "0.1.0"
description = ""
authors = [{ name = "ashkn", email = "ash@ashn.com" }]
package-mode = false
requires-python = ">=3.12"
dependencies = ["httpx (>=0.28.1,<0.29.0)", "schedule (>=1.2.2,<2.0.0)", "python-dotenv (>=1.0.1,<2.0.0)", "python-telegram-bot (>=21.10,<22.0)"]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
